Revert "docs: add builder.io OSS header bar (#2430)"
This reverts commit 480f3bc23b83d87aa2c7703059839e88fd33da5a.
