+++ packages/docs/src/repl/repl-output-panel.tsx
@@ -1,3 +1,4 @@
+import { mutable } from '@builder.io/qwik';
 import { CodeBlock } from '../components/code-block/code-block';
 import { ReplOutputModules } from './repl-output-modules';
 import { ReplOutputSymbols } from './repl-output-symbols';
