+++ packages/docs/src/routes/tutorial/layout!.tsx
@@ -70,14 +70,15 @@ export default component$(() => {
                   </a>
                 </p>
               ) : null}
-              <a
-                class="edit-tutorial"
-                href={`https://github.com/BuilderIO/qwik/edit/main/packages/docs/src/routes/tutorial/${store.appId}`}
-                target="_blank"
-              >
-                <EditIcon width={16} height={16} />
-                <span>Edit Tutorial</span>
-              </a>
+              <p class="edit-tutorial">
+                <a
+                  href={`https://github.com/BuilderIO/qwik/edit/main/packages/docs/src/routes/tutorial/${store.appId}`}
+                  target="_blank"
+                >
+                  <EditIcon width={16} height={16} />
+                  <span>Edit Tutorial</span>
+                </a>
+              </p>
             </div>
           </div>
