+++ packages/docs/src/routes/tutorial/qrl/closures/problem/app.tsx
@@ -10,7 +10,7 @@ export default component$(() => {
           const input = event.target as HTMLInputElement;
           store.name = input.value;
         })}
-        onChange$={$(async () => {
+        onChange$={$(async (event: KeyboardEvent) => {
           if (store.name) alert(store.name);
         })}
         value={store.name}
