+++ tsconfig.json
@@ -79,7 +79,7 @@
   },
   "include": [
     "packages/create-qwik",
-    "packages/docs",
+    // "packages/docs",
     "packages/qwik/src",
     "packages/qwik-city",
     "packages/qwik-react/src",
